Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,20 +2,20 @@
 int main(){
     double x,y;
     char z;
-    scanf("%lf%lf",&x,&y);
+    scanf("%d%d",&x,&y);
     scanf(" %c",&z);
     switch(z){
         case '+':
-        printf("%d",int(x+y));
+        printf("%d",x+y);
         break;
         case '-':
-        printf("%d",int(x-y));
+        printf("%d",x-y);
         break;
         case '*':
-        printf("%d",int(x*y));
+        printf("%d",x*y);
         break;
         case '/':
-        printf("%d",int(x/y));
+        printf("%d",x/y);
         break;
     }
     return 0;
